requirements.txt
fastapi
uvicorn
python-multipart
twilio
langchain-openai
langchain-core
python-dotenv
openai
